🏗️  Changes the folders to make more sense
