+++ services/todoist/auth.py
@@ -2,7 +2,7 @@ import requests
 from todoist import TodoistAPI as SyncTodoistAPI
 
 from config.base import config
-from tasks.todoist_initial_sync import todoist_import_all
+from services.todoist import initial_sync
 
 
 async def todoist_oauth_flow_step_2(code: str, full_sync=False) :
@@ -32,6 +32,6 @@ async def retrieve_user_info(token: str, full_sync=False) -> dict:
 
     if full_sync:
         config.logger.info("Full sync requested")
-        await todoist_import_all(api.state)
+        await initial_sync.todoist_import_all(api.state)
 
     return api.state["user"] | {"token": token}
